+++ src/survival/mod.rs
@@ -1542,11 +1542,13 @@ mod tests {
                 event,
                 ObsRecord::DiscreteState {
                     time: 1.0,
+                    raw_time: 1.0,
                     state: 0,
                     cmt: 3,
                 },
                 ObsRecord::Count {
                     time: 2.0,
+                    raw_time: 2.0,
                     count: 4,
                     cmt: 4,
                 },
@@ -1587,6 +1589,7 @@ mod tests {
             1,
             ObsRecord::DiscreteState {
                 time: 3.0,
+                raw_time: 3.0,
                 state: 1,
                 cmt: 3,
             },
@@ -1620,6 +1623,7 @@ mod tests {
             1,
             ObsRecord::Count {
                 time: 3.0,
+                raw_time: 3.0,
                 count: 2,
                 cmt: 4,
             },